+++ logic/pulsed/pulsed_master_logic.py
@@ -69,7 +69,7 @@ class PulsedMasterLogic(GenericLogic):
 
     # signals for master module (i.e. GUI) coming from PulsedMeasurementLogic
     sigMeasurementDataUpdated = QtCore.Signal()
-    sigTimerUpdated = QtCore.Signal(float, int, float)
+    sigTimerUpdated = QtCore.Signal(float, int, float, float)
     sigFitUpdated = QtCore.Signal(str, np.ndarray, object, str)
     sigMeasurementStatusUpdated = QtCore.Signal(bool, bool)
     sigPulserRunningUpdated = QtCore.Signal(bool)
@@ -324,6 +324,10 @@ class PulsedMasterLogic(GenericLogic):
     @property
     def elapsed_sweeps(self):
         return self.pulsedmeasurementlogic().elapsed_sweeps
+
+    @property
+    def trigger_ratio(self):
+        return self.pulsedmeasurementlogic().trigger_ratio
 
     @property
     def elapsed_time(self):